% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateDiscreteDistributions.R
\name{BivLSD_Cor}
\alias{BivLSD_Cor}
\title{Computes the correlation of the components of a bivariate vector following the
bivariate  logarithmic series distribution}
\usage{
BivLSD_Cor(p1, p2)
}
\arguments{
\item{p1}{parameter \eqn{p_1} of the bivariate  logarithmic series
distribution}

\item{p2}{parameter \eqn{p_2} of the bivariate  logarithmic series
distribution}
}
\value{
Correlation of the components of a bivariate vector following the
 bivariate  logarithmic series distribution
}
\description{
Computes the correlation of the components of a bivariate vector following the
bivariate  logarithmic series distribution
}
