% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateDiscreteDistributions.R
\name{BivLSD_Cov}
\alias{BivLSD_Cov}
\title{Computes the covariance of the components of a bivariate vector following the
bivariate  logarithmic series distribution}
\usage{
BivLSD_Cov(p1, p2)
}
\arguments{
\item{p1}{parameter \eqn{p_1} of the bivariate  logarithmic series
distribution}

\item{p2}{parameter \eqn{p_2} of the bivariate  logarithmic series
distribution}
}
\value{
Covariance of the components of a bivariate vector following the
 bivariate  logarithmic series distribution
}
\description{
Computes the covariance of the components of a bivariate vector following the
bivariate  logarithmic series distribution
}
\examples{
BivLSD_Cov(0.2,0.5)
}
