% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateDiscreteDistributions.R
\name{BivModLSD_Cor}
\alias{BivModLSD_Cor}
\title{Computes the correlation of the components of a bivariate vector following the
bivariate modified  logarithmic series distribution}
\usage{
BivModLSD_Cor(delta, p1, p2)
}
\arguments{
\item{delta}{parameter \eqn{\delta} of the bivariate modified   logarithmic
series distribution}

\item{p1}{parameter \eqn{p_1} of the bivariate modified   logarithmic series
distribution}

\item{p2}{parameter \eqn{p_2} of the bivariate  modified  logarithmic series
distribution}
}
\value{
Covariance of the components of a bivariate vector following the
 bivariate  modified  logarithmic series distribution
}
\description{
Computes the correlation of the components of a bivariate vector following the
bivariate modified  logarithmic series distribution
}
\examples{
BivModLSD_Cor(0.2, 0.3, 0.5)
}
