% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTrawlFunctions.R
\name{fit_Exptrawl}
\alias{fit_Exptrawl}
\title{Fits an exponential trawl function to equidistant time series data}
\usage{
fit_Exptrawl(x, Delta = 1, plotacf = FALSE, lags = 100)
}
\arguments{
\item{x}{vector of equidistant time series data}

\item{Delta}{interval length of the time grid used in the time series, the
default is 1}

\item{plotacf}{binary variable specifying whether or not the empirical and
fitted autocorrelation function should be plotted}

\item{lags}{number of lags to be used in the plot of the autocorrelation
function}
}
\value{
lambda: the memory parameter \eqn{\lambda} in the exponential trawl

LM: the Lebesgue measure of the trawl set associated with the
 exponential trawl,  i.e. \eqn{1/\lambda}.
}
\description{
Fits an exponential trawl function to equidistant time series data
}
\details{
The trawl function is parametrised by the parameter \eqn{\lambda > 0}
 as follows: \deqn{g(x) = e^{\lambda x},  \mbox{ for }  x \le 0.} The
 Lebesgue measure of the corresponding trawl set is given by \eqn{1/\lambda}.
}
