% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMarginalDistributions.R
\name{fit_marginalNB}
\alias{fit_marginalNB}
\title{Fist a negative binomial distribution as marginal law}
\usage{
fit_marginalNB(x, LM, plotdiag = FALSE)
}
\arguments{
\item{x}{vector of equidistant time series data}

\item{LM}{Lebesgue measure of the estimated trawl}

\item{plotdiag}{binary variable specifying whether or not diagnostic plots
should be provided}
}
\value{
m: parameter in the negative binomial marginal distribution

theta: parameter in the negative binomial marginal distribution

a: Here \eqn{a=\theta/(1-\theta)}. This is given for an alternative
 parametrisation of the  negative binomial marginal distribution
}
\description{
Fist a negative binomial distribution as marginal law
}
\details{
The moment estimator for the parameters of the negative binomial
 distribution are given by \deqn{\hat \theta = 1-\mbox{E}(X)/\mbox{Var}(X),}
 and \deqn{\hat m = \mbox{E}(X)(1-\hat \theta)/(\widehat{ \mbox{LM}} \hat
 \theta).}
}
