% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateDiscreteDistributions.R
\name{LSD_Var}
\alias{LSD_Var}
\title{Computes the variance of the logarithmic series distribution}
\usage{
LSD_Var(p)
}
\arguments{
\item{p}{parameter of the logarithmic series distribution}
}
\value{
Variance of the logarithmic series distribution
}
\description{
Computes the variance of the logarithmic series distribution
}
\details{
A random variable \eqn{X} has logarithmic series distribution with
 parameter \eqn{0<p<1}  if \deqn{P(X=x)=(-1)/(\log(1-p))p^x/x, \mbox{ for }
 x=1,2,\dots.}
}
