% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTrawl.R
\name{plot_2and1hist}
\alias{plot_2and1hist}
\title{Plots the bivariate histogram of two time series together with the univariate
histograms}
\usage{
plot_2and1hist(x, y)
}
\arguments{
\item{x}{vector of equidistant time series data}

\item{y}{vector of equidistant time series data (of the same length as x)}
}
\value{
no return value
}
\description{
Plots the bivariate histogram of two time series together with the univariate
histograms
}
\details{
This function plots the bivariate histogram of two time series
 together with the univariate histograms
}
