% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTrawl.R
\name{plot_2and1hist_gg}
\alias{plot_2and1hist_gg}
\title{Plots the bivariate histogram of two time series together with the univariate
histograms using ggplot2}
\usage{
plot_2and1hist_gg(x, y, bivbins = 50, xbins = 30, ybins = 30)
}
\arguments{
\item{x}{vector of equidistant time series data}

\item{y}{vector of equidistant time series data (of the same length as x)}

\item{bivbins}{number of bins in the bivariate histogram}

\item{xbins}{number of bins in the histogram of x}

\item{ybins}{number of bins in the histogram of y}
}
\value{
no return value
}
\description{
Plots the bivariate histogram of two time series together with the univariate
histograms using ggplot2
}
\details{
This function plots the bivariate histogram of two time series
 together with the univariate histograms
}
